DIG-465 BH: Adds extra logging.

+++ docroot/modules/custom/bos_content/modules/node_buildinghousing/src/EventSubscriber/SalesforceBuildingHousingUpdateSubscriber.php
@@ -101,6 +101,16 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
           $bh_meeting = $trigger_event->getEntity();
           $sf_data = $trigger_event->getMappedObject()->getSalesforceRecord();
 
+          // If the settings form has disabled automated updates, then stop.
+          $config = \Drupal::config('node_buildinghousing.settings');
+          if (!str_contains(\Drupal::request()->getRequestUri(), "admin/config/salesforce/boston")
+            && $config->get('pause_auto')) {
+            $trigger_event->disallowPull();
+            return;
+          }
+          $log = $config->get("log_actions");
+          $log && BuildingHousingUtils::log("cleanup", "    PROCESSING Community Meeting event {$sf_data->field("Title__c")}.\n");
+
           if ($supportedLanguages = $sf_data->field('Languages_supported__c')) {
             $supportedLanguages = explode(';', $supportedLanguages);
             $supportedLanguages = implode(', ', $supportedLanguages);
@@ -138,6 +148,16 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
         $bh_project = $trigger_event->getEntity();
         $sf_data = $trigger_event->getMappedObject()->getSalesforceRecord();
         $client = \Drupal::service('salesforce.client');
+
+        // If the settings form has disabled automated updates, then stop.
+        $config = \Drupal::config('node_buildinghousing.settings');
+        if (!str_contains(\Drupal::request()->getRequestUri(), "admin/config/salesforce/boston")
+          && $config->get('pause_auto')) {
+          $trigger_event->disallowPull();
+          return;
+        }
+        $log = $config->get("log_actions");
+        $log && BuildingHousingUtils::log("cleanup", "PROCESSING Project {$sf_data->field("Name")} ({$sf_data->field("Id")}).\n");
 
         // $bh_project->set()
         try {
@@ -200,12 +220,15 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
           return;
         }
 
-      $config = \Drupal::config('node_buildinghousing.settings');
-      if (!str_contains(\Drupal::request()->getRequestUri(), "admin/config/salesforce/boston")
-        && $config->get('pause_auto')) {
-        $trigger_event->disallowPull();
-        return;
-      }
+        // If the settings form has disabled automated updates, then stop.
+        $config = \Drupal::config('node_buildinghousing.settings');
+        if (!str_contains(\Drupal::request()->getRequestUri(), "admin/config/salesforce/boston")
+          && $config->get('pause_auto')) {
+          $trigger_event->disallowPull();
+          return;
+        }
+        $log = $config->get("log_actions");
+        $log && BuildingHousingUtils::log("cleanup", "  PROCESSING Website Update {$sf_data->field("Name")} ({$sf_data->field("Id")}).\n");
 
         // This is a new or updated bh_record.
         // Validate all URL-based fields in the object
@@ -241,7 +264,6 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
             // not fatal.
           }
         }
-
 
         // Read and process the Attachments
         $attachments = [];
@@ -340,6 +362,9 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
 
   private function processAttachments($bh_update, $attachments, $delAttachments) {
 
+    $config = \Drupal::config('node_buildinghousing.settings');
+    $log = $config->get("log_actions");
+
     if (empty($attachments) || empty($bh_update)) {
       return;
     }
@@ -356,6 +381,10 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
 
     $save = FALSE;
     $projectName = basename($bh_project->toUrl()->toString()) ??  'unknown';
+    $count_valid = 0;
+    $count_new = 0;
+    $count_update = 0;
+
     foreach ($attachments as $key => $attachment) {
 
         $storageDirPath = "public://buildinghousing/project/{$projectName}/attachment/{$attachment["fileType"]}/";
@@ -367,6 +396,8 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
         }
         $destination = "{$storageDirPath}{$attachment["sf_id"]}.{$attachment["fileExtension"]}";
 
+        $count_valid++ ;
+
         if (!file_exists($destination)) {
           // New file, or updated file version, so save it.
           if (!$file_data = $this->downloadAttachment($attachment["sf_download_url"])) {
@@ -374,6 +405,10 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
           }
           if (!$file = $this->saveAttachment($file_data, $destination, $attachment)) {
             continue;
+          }
+          else {
+            $count_new++;
+            $log && BuildingHousingUtils::log("cleanup", "    CREATE FILE OBJECT {$attachment["fileName"]}.\n");
           }
         }
         else {
@@ -388,10 +423,14 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
             if (!$file = $this->saveAttachment($file_data, $destination, $attachment)) {
               continue;
             }
+            else {
+              $log && BuildingHousingUtils::log("cleanup", "    UPDATE FILE OBJECT {$attachment["fileName"]}.\n");
+              $count_update++;
+            }
           }
         }
         // Now make a link between the file object amd the bh_ objects.
-        if ($this->linkAttachment($attachment, $file, [$bh_project, $bh_update]) && !$save) {
+        if ($this->linkAttachment($attachment, $file, [$bh_project, $bh_update])) {
           $save = TRUE;
         }
         unset($existing_attachments[$file->id()]);
@@ -399,11 +438,15 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
     }
 
     // Remove any deleted files.
-    if ($delAttachments) {
+    $count_delete = count($existing_attachments);
+    if ($delAttachments && $count_delete > 0) {
       foreach ($existing_attachments as $existing_file) {
+        $log && BuildingHousingUtils::log("cleanup", "    DELETE FILE OBJECT {$existing_file->filename->value}.\n");
         $existing_file->delete();
       }
     }
+
+    $log && BuildingHousingUtils::log("cleanup", "    PROCESSED {$count_valid} File Attachments. {$count_new} Added: {$count_update} Updated: {$count_delete} Deleted.\n");
 
     if ($save) {
       // $bh_update gets automatically saved (later on) by the calling process.
@@ -622,6 +665,9 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
       $currentTextUpdateIds = [];
       $remove = TRUE;
 
+      $config = \Drupal::config('node_buildinghousing.settings');
+      $log = $config->get("log_actions");
+
       // Cache existing text updates in bh_update record for this property.
       foreach ($bh_update->field_bh_text_updates as $key => $currentTextUpdate) {
         $textData = $currentTextUpdate->getValue();
@@ -630,10 +676,15 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
       }
 
       $allowed_chatters = ["TextPost", "ContentPost"];
+      $count_valid = 0;
+      $count_new = 0;
+      $count_update = 0;
+
       // Process the Chatter messages provided.
       foreach ($textUpdateData as $post) {
         if (in_array($post["type"], $allowed_chatters)) {
           if (!empty($post["body"]["text"])) {
+            $count_valid++;
             $message = (!empty($post["body"]["isRichText"]) ? $post["body"]["messageSegments"] : $post["body"]["text"]);
             $drupalPost = [
               'text' => $this->_reformatTextMessage($message, $post["body"]["isRichText"] ?? FALSE),
@@ -645,6 +696,7 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
             if (!array_key_exists($post["id"], $currentTextUpdateIds)) {
               // New posts.
               $bh_update->field_bh_text_updates->appendItem(json_encode($drupalPost));
+              $count_new++;
             }
             else {
               $key = $currentTextUpdateIds[$post["id"]];
@@ -652,9 +704,11 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
               if (!empty($textData->updated) && strtotime($drupalPost["updated"]) != strtotime($textData->updated)) {
                 // Updated posts.
                 $bh_update->field_bh_text_updates->set($key, json_encode($drupalPost));
+                $count_update++;
               }
               if (empty($textData->updated) && strtotime($drupalPost["updated"]) != strtotime($textData->date)) {
                 $bh_update->field_bh_text_updates->set($key, json_encode($drupalPost));
+                $count_update++;
               }
             }
             if ($remove && array_key_exists($post["id"], $currentTextUpdateIds)) {
@@ -663,7 +717,8 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
           }
         }
       }
-      if ($remove) {
+      $count_delete = count($currentTextUpdateIds);
+      if ($remove && $count_delete > 0) {
         // Now remove any chatter items that have been deleted in SF.
         $currentTextUpdateIds = array_flip($currentTextUpdateIds);
         $currentTextUpdateIds = array_reverse($currentTextUpdateIds, TRUE);
@@ -672,6 +727,9 @@ class SalesforceBuildingHousingUpdateSubscriber implements EventSubscriberInterf
           $bh_update->field_bh_text_updates->removeItem($key);
         }
       }
+
+      $log && BuildingHousingUtils::log("cleanup", "    PROCESSED {$count_valid} text messages. {$count_new} Added: {$count_update} Updated: {$count_delete} Deleted.\n");
+
     }
 
     catch (\Exception $e) {
